aws: add instance_ids to run on existing instances

+++ os_tests/libs/resources_aws.py
@@ -94,6 +94,12 @@ class EC2VM(VMResource):
         # enable_efa is option to enable or disable efa when create vms
         # if vm does not support efa, it will be disabled
         self.is_created = False
+        if self.id and not self.params.get("is_allow_delete"):
+            LOG.info(f"Use exists {self.id} which is_allow_delete is {self.params.get("is_allow_delete")}")
+            self.ec2_instance = self.resource.Instance(self.id)
+            self.is_created = True
+            self.instance_type = self.ec2_instance.instance_type
+            return True
 
         try:
             self.efa_support = self.client.describe_instance_types(
@@ -476,6 +482,9 @@ class EC2VM(VMResource):
             return False
 
     def delete(self, wait=True, loops=4):
+        if self.id and not self.params.get("is_allow_delete"):
+            LOG.info(f"Use exists {self.id} which is_allow_delete is {self.params.get("is_allow_delete")}")
+            return True
         try:
             LOG.info("Deleting instance: %s" % self.ec2_instance.id)
             self.ec2_instance.terminate()
